[Table] cleanup code, extract types and sub components
