+++ src/app/components/Table/index.module.scss
@@ -1,35 +1,8 @@
 .headRow {
   // Using ::before and ::after to enable border radii on odd rows, even when the first-/last-child is hidden
-  &::before {
-    @apply tw-table-cell tw-sticky tw-top-0 tw-z-10  tw-w-4 tw-bg-gray-1;
-    content: '';
-  }
-
+  &::before,
   &::after {
-    @apply tw-table-cell tw-sticky tw-top-0 tw-z-10  tw-w-2 tw-bg-gray-1;
+    @apply tw-table-cell tw-sticky tw-top-0 tw-z-10 tw-w-2 tw-bg-gray-1;
     content: '';
-  }
-}
-
-.row {
-  @apply tw-cursor-pointer sm:tw-cursor-auto;
-
-  // Using ::before and ::after to enable border radii on odd rows, even when the first-/last-child is hidden
-  &::before {
-    @apply tw-table-cell tw-w-4 tw-rounded-l-lg;
-    content: '';
-  }
-
-  &::after {
-    @apply tw-table-cell tw-w-2 tw-rounded-r-lg;
-    content: '';
-  }
-
-  &:nth-child(odd) {
-    & > *,
-    &::before,
-    &::after {
-      @apply tw-bg-gray-2;
-    }
   }
 }
