+++ src/app/components/Table/index.stories.tsx
@@ -36,7 +36,7 @@ const rowsSmall = [
 const rowsLarge = new Array(20).fill(null).map((e, index) => ({
   key: index,
   name: String.fromCharCode('a'.charCodeAt(0) + index),
-  source: '0x959919178CEff4a7bf5A1EdCEFFe2Eb965709d97',
+  source: '0x653c65d44D062344A35a0204f94C1D0a59b54a25',
   target: '0xc37A85e35d7eECC82c4544dcba84CF7E61e1F1a3',
   progress: Math.random(),
   flag: Math.random() > 0.5,
