+++ src/styles/__tests__/media.test.ts
@@ -1,11 +1,12 @@
-import { media, sizes } from '../media';
+import { media } from '../media';
 import { css } from 'styled-components/macro';
+import { BreakpointWidths } from '../../types';
 
 describe('media', () => {
   it('should return media query in css', () => {
     const mediaQuery = media.sm`color:red;`.join('');
     const cssVersion = css`
-      @media (min-width: ${sizes.sm}px) {
+      @media (min-width: ${BreakpointWidths.sm}px) {
         color: red;
       }
     `.join('');
