+++ src/styles/media.ts
@@ -1,3 +1,4 @@
+import { BreakpointWidths } from '../types';
 /*
  * Media queries utility
  */
@@ -11,32 +12,18 @@ import {
   Interpolation,
   FlattenInterpolation,
 } from 'styled-components/macro';
-
-/*
- * Taken from https://github.com/DefinitelyTyped/DefinitelyTyped/issues/32914
- */
-
-// Update your breakpoints if you want
-export const sizes = {
-  sm: 576,
-  md: 768,
-  lg: 992,
-  xl: 1200,
-};
+import { Breakpoint } from '../types';
 
 // Iterate through the sizes and create a media template
-export const media = (Object.keys(sizes) as Array<keyof typeof sizes>).reduce(
-  (acc, label) => {
-    acc[label] = (first: any, ...interpolations: any[]) => css`
-      @media (min-width: ${sizes[label]}px) {
-        ${css(first, ...interpolations)}
-      }
-    `;
+export const media = Object.keys(BreakpointWidths).reduce((acc, label) => {
+  acc[label] = (first: any, ...interpolations: any[]) => css`
+    @media (min-width: ${BreakpointWidths[label]}px) {
+      ${css(first, ...interpolations)}
+    }
+  `;
 
-    return acc;
-  },
-  {} as { [key in keyof typeof sizes]: MediaFunction },
-);
+  return acc;
+}, {} as { [key in Breakpoint]: MediaFunction });
 
 /*
  * @types/styled-component is not working properly as explained in the github issue referenced above.
